+++ CHANGELOG.md
@@ -5,6 +5,13 @@ Headlines: Added, Changed, Deprecated, Removed, Fixed, Security
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.14] - 2023-02-28
+
+### Fixed
+
+- Client-side validators were overwriting errors passed from the server.
+- Cleared timers after submit is complete.
+
 ## [0.5.13] - 2023-02-28
 
 ### Added